+++ addon/commands/indent-node.ts
@@ -8,8 +8,8 @@ type IndentNodeArgs = {
 };
 
 /**
- * Returns a command to indent the node(s) in the selection. Only nodes with 
- * nodespec `indentable` equal to true and `indentationLevel` as node attribute can be indented.
+ * Returns a command to indent the node(s) in the selection.
+ * Only nodes with `indentationLevel` as node attribute can be indented.
  * @direction the number -1 or 1 to indent back or further
  * @predicate extra check to see if a node should be allowed to indent
  * @maxLevel max level to indent
@@ -25,7 +25,7 @@ export function indentNode({
     const applicableNodes: { node: PNode; pos: number }[] = [];
     state.doc.nodesBetween(from, to, (node, pos, parent) => {
       if (
-        node.type.spec.indentable &&
+        node.attrs.indentationLevel !== undefined &&
         predicate(node, pos, parent) &&
         ((direction === -1 && node.attrs.indentationLevel > 0) ||
           (direction === 1 && node.attrs.indentationLevel < maxLevel))